working on delete for researchers. copied modal forms from Kai.
