+++ simcon_project/simcon_project/urls.py
@@ -15,8 +15,8 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from users.views import RedirectFromLogin, StudentView, ResearcherView, ViewResponse, \
-  UpdateOverallResponseFeedback, UpdateNodeResponseFeedback, SettingsView, TemplateStartView
+from users.views import RedirectFromLogin, StudentView, ResearcherView, ViewResponse, UpdateOverallResponseFeedback, \
+    UpdateNodeResponseFeedback, TemplateStartView, ResearcherUserView, StudentSettingsView, StudentUserView
 from django.conf.urls import include
 from django.contrib.auth import views
 
@@ -30,14 +30,14 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('django.contrib.auth.urls')),
     path('redirect-from-login/', RedirectFromLogin, name="RedirectFromLogin"),
-    path('student-view/', StudentView, name="StudentView"),
     path('researcher-view/', ResearcherView, name="ResearcherView"),
+    path('researcher-view/template-management/', include('conversation_templates.urls'), name="TemplateManagementView"),
     path('user-view/', ResearcherUserView, name="ResearcherUserView"),
+    path('researcher-settings/', include('users.urls'), name="ResearcherSettingsView"),
+    path('student-view/', StudentView, name="StudentView"),
     path('studentuser-view/', StudentUserView, name="StudentUserView"),
-    path('settings/', ResearcherSettingsView, name="ResearcherSettingsView"),
     path('student-settings/', StudentSettingsView, name="StudentSettingsView"),
     path('template-start/<str:name>/', TemplateStartView, name="TemplateStartView"),
-    path('researcher-view/template-management/', include('conversation_templates.urls'), name="TemplateManagementView"),
     path('view-response/', ViewResponse, name="ViewResponse"),
     path('view-response/<uuid:pk>/update/', UpdateOverallResponseFeedback, name='UpdateOverallResponseFeedback'),
     path('view-response/<uuid:pk>/updatenode/', UpdateNodeResponseFeedback, name='UpdateNodeResponseFeedback'),
