+++ simcon_project/templates/settings/settings_buttons_template.html
@@ -1,0 +1,16 @@
+<!--
+    Delete or remove button TemplateColumn template used in
+    FolderTemplateTable and AllTemplateTable
+ -->
+{% block content %}
+{% if in_folder %}
+    <form action="{% url 'settings:remove_template' record.id %}" method="post">
+        {% csrf_token %}
+        <input type="hidden" name="back" value="{{ request.path }}">
+        <button id="remove-template" class="btn btn-danger" type="submit">Remove</button>
+    </form>
+{% else %}
+    <button id="delete-template" class="bs-modal btn btn-danger" type="button" name="button"
+            data-form-url="{% url 'management:delete_template' record.id %}">Delete</button>
+{% endif %}
+{% endblock %}
