+++ simcon_project/templates/student_settings_view.html
@@ -1,5 +1,3 @@
-{% load static %}
-{% load render_table from django_tables2 %}
 {% extends "student_base.html" %}
 {% block nav-studentssettingsView %}{% endblock %}
 {% block content %}
@@ -16,54 +14,15 @@
     <style type="text/css">
     form {display:inline; margin:0px; padding:0px;}
     </style>
-
-
     <title>
     {% block title %}Simulated Conversations | Settings{% endblock %}
     </title>
-
     {% block extrahead %} {% endblock %}
   </head>
   <body>
     <div class="container">
       <h1> This is the Settings Page</h1>
     </div>
-        {% block content %}
-          {% if messages %}
-          <ul class="messages">
-              {% for message in messages %}
-              <li {% if message.tags %} class="{{ message.tags }}"{% endif %}> {{ message }}</li>
-              {% endfor %}
-          </ul>
-          {% endif %}
-          <form id="change_password_form" method="post">
-            {% csrf_token %}
-            {{ change_password_form }}
-            <button type="submit" form="change_password_form">Save changes</button>
-          </form>
-          <br>
-          {% if delete_researchers_form %}
-            <form id="delete_researchers_form" method="post">
-              {% csrf_token %}
-              {% render_table researchers_table %}
-              <button type="submit" form="delete_researchers_form">Delete Selected Researchers</button>
-            </form>
-          {% endif %}
-          <br>
-          {% if add_researcher_form %}
-            <form id="add_researcher_form" method="post">
-              {% csrf_token %}
-              {{ add_researcher_form }}
-              <button type="submit" form="add_researcher_form">Send Registration Email</button>
-            </form>
-          {% endif %}
-        {% endblock %}
-    </div> <!-- /container -->
-
-    <!-- Bootstrap core JavaScript
-    ================================================== -->
-    <!-- Placed at the end of the document so the pages load faster -->
   </body>
 </html>
 {% endblock %}
-
