+++ simcon_project/users/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from .views import *
+
+app_name = 'settings'
+urlpatterns = [
+    path('', ResearcherSettingsView, name="main"),
+    path('delete/<str:email>', ResearcherDeleteView.as_view(), name="delete_researcher"),
+]
