+++ simcon_project/users/views/__init__.py
@@ -1,6 +1,6 @@
 from .redirect_from_login import RedirectFromLogin
 from .researcher_home import ResearcherView
-from .researcher_settings_page import ResearcherSettingsView
+from .researcher_settings_page import ResearcherSettingsView, ResearcherDeleteView
 from .researcher_user_page import ResearcherUserView
 from .studentuser_page import StudentUserView
 from .student_settings_page import StudentSettingsView
